+++ src/main/kotlin/me/rafaelldi/aspire/run/AspireHostConfigurationViewModel.kt
@@ -16,6 +16,7 @@ class AspireHostConfigurationViewModel(
     private val lifetime: Lifetime,
     private val runnableProjectsModel: RunnableProjectsModel?,
     val projectSelector: ProjectSelector,
+    val environmentVariablesEditor: EnvironmentVariablesEditor,
     separator: ViewSeparator,
     val urlEditor: TextEditor,
     val dotNetBrowserSettingsEditor: BrowserSettingsEditor
@@ -23,12 +24,15 @@ class AspireHostConfigurationViewModel(
     override val controls: List<ControlBase> =
         listOf(
             projectSelector,
+            environmentVariablesEditor,
             separator,
             urlEditor,
             dotNetBrowserSettingsEditor
         )
 
     private var isLoaded = false
+
+    var trackEnvs = true
     var trackUrl = true
 
     init {
@@ -44,6 +48,7 @@ class AspireHostConfigurationViewModel(
             )
         }
 
+        environmentVariablesEditor.envs.advise(lifetime) { handleEnvValueChange() }
         urlEditor.text.advise(lifetime) { handleUrlValueChange() }
     }
 
@@ -53,11 +58,20 @@ class AspireHostConfigurationViewModel(
         }
 
         val runOptions = project.getRunOptions()
+        environmentVariablesEditor.envs.set(runOptions.environmentVariables)
         val startBrowserUrl = runOptions.startBrowserUrl
         if (startBrowserUrl.isNotEmpty()) {
             urlEditor.defaultValue.value = startBrowserUrl
             urlEditor.text.value = startBrowserUrl
             dotNetBrowserSettingsEditor.settings.value = BrowserSettings(runOptions.launchBrowser, false, null)
+        }
+    }
+
+    private fun handleEnvValueChange() {
+        projectSelector.project.valueOrNull?.let {
+            val envs = it.environmentVariables.associate { pair -> pair.key to pair.value }.toSortedMap()
+            val editorEnvs = environmentVariablesEditor.envs.value.toSortedMap()
+            trackEnvs = envs == editorEnvs
         }
     }
 
@@ -68,9 +82,16 @@ class AspireHostConfigurationViewModel(
         }
     }
 
-    fun reset(projectFilePath: String, trackUrl: Boolean, dotNetStartBrowserParameters: DotNetStartBrowserParameters) {
+    fun reset(
+        projectFilePath: String,
+        trackEnvs: Boolean,
+        envs: Map<String, String>,
+        trackUrl: Boolean,
+        dotNetStartBrowserParameters: DotNetStartBrowserParameters
+    ) {
         isLoaded = false
 
+        this.trackEnvs = trackEnvs
         this.trackUrl = trackUrl
 
         runnableProjectsModel?.projects?.adviseOnce(lifetime) { projectList ->
@@ -115,6 +136,10 @@ class AspireHostConfigurationViewModel(
                     it.projectFilePath == projectFilePath && it.kind == AspireRunnableProjectKinds.AspireHost
                 }?.let { project ->
                     projectSelector.project.set(project)
+
+                    val effectiveEnvs =
+                        if (trackEnvs) project.environmentVariables.associate { it.key to it.value } else envs
+                    environmentVariablesEditor.envs.set(effectiveEnvs)
 
                     val runOptions = project.getRunOptions()
                     val effectiveUrl = if (trackUrl) runOptions.startBrowserUrl else dotNetStartBrowserParameters.url